requirements.txt
asgiref==3.5.0
Django==4.0.2
numpy
Pillow
pytesseract
sqlparse
tzdata
